% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aux_CalculateCellInteractions.R
\name{ActualGrowth_CrushableDistanceConstraint}
\alias{ActualGrowth_CrushableDistanceConstraint}
\title{ActualGrowth_CrushableDistanceConstraint}
\usage{
ActualGrowth_CrushableDistanceConstraint(n)
}
\arguments{
\item{n}{The cell number for which to calculate the crushable distance.}
}
\value{
None
}
\description{
This function calculates the actual growth as constrained by nearby cells that are "incompressible" based on a "maximum non-crushable distance".
}
\details{
This function calculates the actual growth rates of the cell, taking into account the constraints imposed by neighboring cells that are "incompressible". It checks if the lower radial growth rate or the upper tangential growth rate of the cell violate the maximum non-crushable distances. If they do, the growth rates are adjusted accordingly.

code translated from Pascal to R by: ChatGPT, Annemarie Eckes-Shephard
}
\author{
David Drew
}
\keyword{ActualGrowth_CrushableDistanceConstraint}
