% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aux_CalculateCellInteractions.R
\name{ActualGrowth_RadTanRatioConstraint}
\alias{ActualGrowth_RadTanRatioConstraint}
\title{ActualGrowth_RadTanRatioConstraint}
\usage{
ActualGrowth_RadTanRatioConstraint(n)
}
\value{
None
}
\description{
This function calculates the actual growth rates of the cell, taking into account the radial-tangential ratio constraint. It ensures that the ratio of the sum of tangential growth rates to the sum of radial growth rates does not exceed a specified threshold.
}
\details{
This function checks if the current radial and tangential growth rates, along with their upper and lower limits, violate the radial-tangential ratio constraint. If the constraint is violated, the growth rates are adjusted accordingly to satisfy the constraint.

code translated from Pascal to R by: ChatGPT, Annemarie Eckes-Shephard
}
\author{
David Drew
}
\keyword{ActualGrowth_RadTanRatioConstraint}
