% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aux_functions.R
\name{Apoptosis}
\alias{Apoptosis}
\title{Perform cell apoptosis assessment}
\usage{
Apoptosis(n, day)
}
\arguments{
\item{n}{cellnumber  - index of the cell to evaluate}

\item{day}{simulation day}
}
\value{
None (modifies input vectors in-place)

DaysSinceSecThickening A vector containing the number of days since secondary thickening for each cell

CritThickeningTimeThreshold A vector containing the critical thickening time threshold for each cell

JustDiedFlag A vector indicating whether a cell has just died

DaysSinceDeath A vector containing the number of days since death for each cell

WallThickRate A vector containing the wall thickening rate for each cell

CellDead A vector indicating whether a cell is dead

DifferentiationStatus A vector indicating the differentiation status of each cell

RadPosition A vector containing the radial position of each cell

MaxDifferentiatedCellPosition A vector containing the maximum radial position of differentiated cells for each day

MaxDeadCellPosition The maximum radial position of dead cells
}
\description{
This function assesses whether a cell must undergo apoptosis based on the criteria specified in the Pascal procedure.
}
\details{
This function evaluates whether a cell must undergo apoptosis by comparing the number of days since secondary thickening
with the critical thickening time threshold. If the cell has been thickening or differentiating for a set maximum period,
it is marked as dead. The relevant vectors are updated accordingly.

Code translated from Pascal to R by: ChatGPT, Annemarie Eckes-Shephard
}
\author{
David Drew
}
\keyword{apoptosis,}
\keyword{cell,}
\keyword{differentiation}
