% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aux_functions.R
\name{AuxinField}
\alias{AuxinField}
\title{AuxinField}
\usage{
AuxinField(day, n)
}
\description{
Calculates the auxin field for each cell based on specific parameters.
}
\details{
This function calculates the auxin concentration for each cell using
         the given parameters and formulas. It considers factors such as leaf
         modifiers, environmental conditions, height, stem activity, and the
         distance from the initial cell. The resulting auxin concentration is
         stored in the AuxinConc vector for each cell.

Code translated from Pascal to R by: ChatGPT, Annemarie Eckes-Shephard
}
\author{
David Drew
}
\keyword{AuxinField,}
\keyword{activity,}
\keyword{auxin}
\keyword{cell}
\keyword{cell,}
\keyword{concentration,}
\keyword{condition,}
\keyword{distance,}
\keyword{environmental}
\keyword{height,}
\keyword{initial}
\keyword{leaf}
\keyword{modifier,}
\keyword{stem}
