% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildCellPopulation.R
\name{BuildCellPopulation}
\alias{BuildCellPopulation}
\title{BuildCellPopulation}
\usage{
BuildCellPopulation()
}
\value{
None
}
\description{
Build an initial population of homogeneous cells which will form the basis for the modeling population.
}
\details{
This function builds an initial population of cells for modeling purposes. It assigns formation dates, cell types, positions, and other properties to each cell.
The population is based on user-specified parameters and calculations.

Code translated from Pascal to R by: ChatGPT, Annemarie Eckes-Shephard
}
\author{
David Drew
}
\keyword{parameters}
\keyword{parameters,}
\keyword{reading}
\keyword{xylem}
