% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aux_CalculateCellInteractions.R
\name{CalculateActualSize}
\alias{CalculateActualSize}
\title{Calculate Actual Size}
\usage{
CalculateActualSize(n, positioncellcounter)
}
\arguments{
\item{positioncellcounter}{passed on function CrushCells.}

\item{cellnumber}{The cell number to calculate the actual size for.}
}
\value{
NULL (The function updates the provided arrays by reference.)
}
\description{
This function calculates the actual size of cells based on growth rates.
}
\details{
The function updates the values of cellRD, cellTD, RadPosition, TanPosition, CellLength, CellVolume, and CellCSArea for each cell.

code translated from Pascal to R by: ChatGPT, Annemarie Eckes-Shephard
}
\author{
David Drew
}
\keyword{"Calculate}
\keyword{Actual}
\keyword{Size",}
\keyword{cell}
\keyword{growth,}
\keyword{size}
