% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateCellGrowthPotential.R
\name{CalculateCellGrowthPotential}
\alias{CalculateCellGrowthPotential}
\title{CalculateCellGrowthPotential}
\usage{
CalculateCellGrowthPotential(day, TotalLoopSize)
}
\description{
Calculates the cell growth potential based on the given climate forcings and parameters
}
\details{
This function calculates the cell growth potential by considering factors such as temperature, water potential, and photosynthate availability.

Code translated from Pascal to R by: ChatGPT, Annemarie Eckes-Shephard
}
\author{
David Drew
}
\keyword{cell}
\keyword{growth}
\keyword{growth,growth}
\keyword{increment,}
\keyword{potential,}
