% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateCellInteractions.R
\name{CalculateCellInteractions}
\alias{CalculateCellInteractions}
\title{Calculate Cell Interactions}
\usage{
CalculateCellInteractions(day)
}
\value{
None
}
\description{
This function calculates the interaction between cells in two dimensions.
}
\details{
Calculate the interaction between cells in two dimensions

code translated from Pascal to R by: ChatGPT, Annemarie Eckes-Shephard
}
\author{
David Drew
}
\keyword{"cell}
\keyword{"two}
\keyword{dimensions"}
\keyword{interactions",}
