% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateCellSpecificAllocation.R
\name{CalculateCellSpecificAllocation}
\alias{CalculateCellSpecificAllocation}
\title{Calculate Cell-Specific Allocation}
\usage{
CalculateCellSpecificAllocation(day)
}
\arguments{
\item{day}{simulation day}
}
\value{
None
}
\description{
This function calculates the amount of photosynthate allocated to the stem that is available to each cell.
}
\details{
The function calculates the number of cells in the stem, the volume of living xylem tissue, the total dry mass of the stem, the volume of dead tissue in the stem, and the average amount of photosynthate that can be allocated to each living cell in the stem.

Code translated from Pascal to R by: ChatGPT, Annemarie Eckes-Shephard
}
\author{
David Drew
}
\keyword{Allocation,}
\keyword{Photosynthate}
\keyword{Stem,}
