% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aux_functs_daily_segment_means.R
\name{CalculateConductiveCells}
\alias{CalculateConductiveCells}
\title{Calculate Conductive Cells}
\usage{
CalculateConductiveCells(n, day)
}
\arguments{
\item{n}{The cell number index}

\item{day}{The log day}
}
\value{
None
}
\description{
This function calculates the conductive cells based on the number of days since their death.
Cells become conductive units after a certain period following their death.
}
