% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aux_functions.R
\name{CalculateDensity}
\alias{CalculateDensity}
\title{CalculateDensity}
\usage{
CalculateDensity()
}
\value{
None
}
\description{
This function calculates the wood density based on the mean fibre radial density
and mean fibre wall thickness of the segment. If the mean values are greater than
0, the wood density is calculated using the specified formulas. Otherwise, the
previous wood density value is used.
}
\details{
This function calculates the wood density based on the mean fibre radial density
and mean fibre wall thickness of the segment. If the mean values are greater than
0, the wood density is calculated using the specified formulas. Otherwise, the
previous wood density value is used.

Code translated from Pascal to R by: ChatGPT, Annemarie Eckes-Shephard
}
\author{
David Drew
}
\keyword{CalculateDensity}
