% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateDevelopmentalDurations.R
\name{CalculateDevelopmentalDurations}
\alias{CalculateDevelopmentalDurations}
\title{Calculate Developmental Durations}
\usage{
CalculateDevelopmentalDurations(day)
}
\value{
The calculated wall thickening duration for the current day.
}
\description{
This function calculates the duration of wall thickening for the current day based on environmental control.
}
\details{
The function calculates the duration of wall thickening for the current day.

code translated from Pascal to R by: ChatGPT, Annemarie Eckes-Shephard
}
\author{
David Drew
}
\keyword{"edit}
\keyword{depending}
\keyword{function}
\keyword{on}
\keyword{the}
\keyword{this}
\keyword{translate"}
\keyword{you}
