% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aux_growth_functions.R
\name{CalculateGrowthIncrement}
\alias{CalculateGrowthIncrement}
\title{CalculateGrowthIncrement}
\usage{
CalculateGrowthIncrement(TotalLoopSize, day, CellNumber)
}
\arguments{
\item{day}{The current day.}

\item{CellNumber}{The number of cells.}
}
\value{
The calculated growth increment.
}
\description{
Calculates the growth increment based on the given parameters.
}
\details{
This function calculates the growth increment for a specific cell based on various factors such as CAMBIALINITIAL, RAYINITIALC, MERISTEMATIC, and SECONDARYTHICKENING.

Code translated from Pascal to R by: ChatGPT, Annemarie Eckes-Shephard
}
\examples{
CalculateGrowthIncrement(CellType = "CAMBIALINITIAL", CellTD = 1, TanPosition = 0.5, RadPosition = 0.2, CellKey = 123, TotalLoopSize = 100, day = 1, CellNumber = 10)
}
\author{
David Drew
}
\keyword{"cell}
\keyword{"growth}
\keyword{growth"}
\keyword{increment",}
