% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aux_functions.R
\name{CalculateInitialActualGrowthRates}
\alias{CalculateInitialActualGrowthRates}
\title{CalculateInitialActualGrowthRates}
\usage{
CalculateInitialActualGrowthRates(
  n,
  PotCellTanGrowthRate,
  PotCellRadGrowthRate
)
}
\arguments{
\item{n}{CellNumber index of the cell that is currently to be treated by this function.}

\item{PotCellTanGrowthrate}{potential tangential growth rate calculated in calling function CalculateCellGrowthPotential()}

\item{PotCellRadGrowthrate}{potential radial growth rate calculated in calling function CalculateCellGrowthPotential()}
}
\description{
Calculates the upper and lower actual growth rates based on the potential radial and tangential growth rates and user-defined proportions.
}
\details{
This function calculates the upper and lower tangential growth rates from the potential radial
and tangential growth rates based on user-defined upper and lower proportions.
If the cell type is "VESSEL," the lower radial growth rate is determined by multiplying
the potential radial growth rate by the lower growth proportion, and the upper radial growth
rate is determined by subtracting the lower radial growth rate from the potential radial growth rate.
 If the cell type is not "VESSEL," the lower radial growth rate is set to 0, and the upper
 radial growth rate remains the same as the potential radial growth rate.
  The lower tangential growth rate is half of the potential tangential growth rate,
  and the upper tangential growth rate is also half of the potential tangential growth rate.

Code translated from Pascal to R by: ChatGPT, Annemarie Eckes-Shephard
}
\author{
David Drew
}
\keyword{"actual}
\keyword{"radial}
\keyword{"tangential}
\keyword{growth}
\keyword{growth"}
\keyword{growth",}
\keyword{rates",}
