% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateModifiers.R
\name{CalculateModifiers}
\alias{CalculateModifiers}
\title{Calculate Modifiers}
\usage{
CalculateModifiers()
}
\value{
None
}
\description{
Calculates various modifiers based on specific conditions and data.
}
\details{
This function calculates modifiers such as fWaterStress, fCytokinin, fTemp,
         FEnvironment, StemActivityModifier, LeafGrowthModifier, LeafModifierAuxin,
         FoliageAllocationMean, and StemAllocationMean.

Code translated from Pascal to R by: ChatGPT, Annemarie Eckes-Shephard
}
\author{
David Drew
}
\keyword{growth}
\keyword{modifiers}
\keyword{xylem}
