% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateNeighbours.R
\name{CalculateNeighbours}
\alias{CalculateNeighbours}
\title{Calculate Neighbours}
\usage{
CalculateNeighbours(TotalLoopSize)
}
\description{
This function calculates the lower and upper radial as well as lower and upper tangential neighbours for each cell.
}
\details{
The function iterates through each cell and determines its neighbours based on radial and tangential positions.

Code translated from Pascal to R by: ChatGPT, Annemarie Eckes-Shephard
}
\author{
David Drew
}
