% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aux_Functions_GrowCellWall.R
\name{CalculatePotentialWallThickRate}
\alias{CalculatePotentialWallThickRate}
\title{Calculate Potential Wall Thick Rate}
\usage{
CalculatePotentialWallThickRate(day, n)
}
\arguments{
\item{n}{The cell number for which to calculate the potential wall thickening rate.}

\item{d}{simulation day}
}
\value{
None (the function updates variables)
}
\description{
Calculates the potential wall thickening rate for each cell.
}
\details{
This function calculates the potential wall thickening rate based on the cell type. If the cell type
is not a ray, ray initial X, or ray initial P, the function calculates the maximum thickening cell position and
then proceeds with the calculation of the wall thickening rate. The potential volumetric wall thickening rate is
calculated as the product of photosynthate per cell for the current day and the wall conversion factor, adjusted
to the appropriate units. The potential volumetric wall thickening rate is then converted to a gain in volume
and subtracted from the lumen volume to obtain the potential lumen volume. If the potential lumen volume is greater
than zero, the potential lumen cross-sectional area is calculated using the body cross-sectional area function.
Finally, the potential area wall thickening rate is obtained by subtracting the potential lumen cross-sectional area
from the current lumen cross-sectional area. If the cell type is a ray, ray initial X, or ray initial P, the
potential area wall thickening rate is set to zero since rays are not allowed to secondary thicken.

Code translated from Pascal to R by: ChatGPT, Annemarie Eckes-Shephard
}
\author{
David Drew
}
\keyword{area}
\keyword{cell}
\keyword{cross-sectional}
\keyword{lumen}
\keyword{potential}
\keyword{rate,}
\keyword{thickening,}
\keyword{type,}
\keyword{volume,}
\keyword{wall}
