% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aux_Functions_GrowCellWall.R
\name{CalculatePrimaryWallThickRate}
\alias{CalculatePrimaryWallThickRate}
\title{Calculate Primary Wall Thick Rate}
\usage{
CalculatePrimaryWallThickRate(n)
}
\arguments{
\item{n}{The cell number for which to calculate the primary wall thickening rate.}
}
\value{
None (the function updates variables)
}
\description{
Calculates the primary wall thickening rate for a cell.
}
\details{
This function calculates the primary wall thickening rate based on the differentiation status
of the cell and the lumen volume. If the cell is not undergoing secondary thickening or does not have
a new secondary wall, the lumen volume is calculated assuming a primary wall thickness of 0.25um.
The cell wall cross-sectional area is updated accordingly, and the primary wall thickening rate is
determined as the difference between the updated cell wall cross-sectional area and the previous value.
If the updated cell wall cross-sectional area is smaller than the previous value, the primary wall
thickening rate is set to 0.

Code translated from Pascal to R by: ChatGPT, Annemarie Eckes-Shephard
}
\author{
David Drew
}
\keyword{differentiation,}
\keyword{lumen}
\keyword{primary}
\keyword{rate,}
\keyword{thickening,}
\keyword{volume}
\keyword{wall,}
