% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aux_CalculateCellInteractions.R
\name{CalculateRadCrushableDistance}
\alias{CalculateRadCrushableDistance}
\title{Calculate the maximum distance pithward that the cell can expand}
\usage{
CalculateRadCrushableDistance(n, positioncellcounter)
}
\arguments{
\item{n}{The cell number for which to calculate the crushable distance.}

\item{positioncellcounter}{The position cell counter.}
}
\value{
The updated MaxRadNonCrushableCellDistanceLower value.
}
\description{
This function calculates the maximum distance pithward that the cell can expand based on certain conditions.
}
\details{
The function checks if the potential growth of the cell would infringe upon another cell's barkward edge. If it does, and the other cell meets certain conditions, the maximum crushable distance is updated.

Code translated from Pascal to R by: ChatGPT, Annemarie Eckes-Shephard
}
\author{
David Drew
}
\keyword{CalculateRadCrushableDistance}
