% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aux_CalculateCellInteractions.R
\name{CalculateRadialShift}
\alias{CalculateRadialShift}
\title{CalculateRadialShift}
\usage{
CalculateRadialShift()
}
\value{
Returns the updated values of NewRadialShift.
}
\description{
Calculate/re-calculate the total growth "behind" each cell to use for individual cell positions.
}
\details{
This function calculates the radial shift of cells based on various conditions and parameters.

Code translated from Pascal to R by: ChatGPT, Annemarie Eckes-Shephard
}
\author{
David Drew
}
\keyword{calculate,}
\keyword{cell}
\keyword{position}
\keyword{radial}
\keyword{shift,}
