% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aux_Functions_GrowCellWall.R
\name{CalculateSecondaryWallThickRate}
\alias{CalculateSecondaryWallThickRate}
\title{Calculate Secondary Wall Thick Rate}
\usage{
CalculateSecondaryWallThickRate(day, n)
}
\arguments{
\item{n}{The cell number for which to calculate the secondary wall thickening rate.}
}
\value{
None (the function updates variables)
}
\description{
Calculates the amount of transverse wall area and wall thickness added to each cell.
}
\details{
This function calculates the secondary wall thickening rate based on the differentiation status
of the cell. If the cell is undergoing secondary thickening, the new secondary wall flag is set to true.
The maximum wall thickening rate is determined based on the user-specified maximum for the species and the
potential area wall thickening rate. If the potential area wall thickening rate exceeds the maximum, the
wall thickening rate is limited to the maximum value. If the potential area wall thickening rate is less
than the minimum wall thickening rate for storage use, the wall thickening rate is set to the minimum value
if there is wall thickening stored mass available; otherwise, it is set to the potential area wall thickening
rate. If the cell is not undergoing secondary thickening, the wall thickening rate is set to 0.

Code translated from Pascal to R by: ChatGPT, Annemarie Eckes-Shephard
}
\author{
David Drew
}
\keyword{differentiation,}
\keyword{maximum}
\keyword{rate,}
\keyword{secondary}
\keyword{storage}
\keyword{thickening,}
\keyword{use}
\keyword{wall,}
