% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aux_Functions_GrowCellWall.R
\name{CalculateStorage}
\alias{CalculateStorage}
\title{CalculateStorage}
\usage{
CalculateStorage(day)
}
\arguments{
\item{day}{simulation day}

\item{n}{The cell number index}
}
\value{
None
}
\description{
Calculates the storage amount and updates the wall thickness storage mass.
}
\details{
This function calculates the total storage amount based on the potential storage increase mass
and the number of cells in the storage. It then updates the wall thickness storage mass.

Code translated from Pascal to R by: ChatGPT, Annemarie Eckes-Shephard
}
\author{
David Drew
}
\keyword{Calculate}
\keyword{cells,}
\keyword{mass,}
\keyword{potential}
\keyword{storage,}
\keyword{thickness,}
\keyword{wall}
