% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aux_Functions_GrowCellWall.R
\name{CalculateStorageChange}
\alias{CalculateStorageChange}
\title{CalculateStorageChange}
\usage{
CalculateStorageChange(day, n)
}
\arguments{
\item{day}{simulation day}

\item{n}{The cell number index}
}
\value{
None
}
\description{
Calculates the change in storage for a given cell.
}
\details{
This function calculates the potential storage increase and decrease volumes and masses for a specific cell based on its properties and status.
The calculated values are used to update the storage variables.

Code translated from Pascal to R by: ChatGPT, Annemarie Eckes-Shephard
}
\author{
David Drew
}
\keyword{Calculate}
\keyword{cell}
\keyword{change,}
\keyword{mass,}
\keyword{potential,}
\keyword{storage}
\keyword{volume,}
