% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aux_CalculateCellInteractions.R
\name{CalculateTanCrushableDistance}
\alias{CalculateTanCrushableDistance}
\title{CalculateTanCrushableDistance}
\usage{
CalculateTanCrushableDistance(n, PositionCellCounter)
}
\arguments{
\item{n}{cell number index}
}
\value{
None
}
\description{
Calculate the maximum distance left and right of the current cell that it can expand tangentially
}
\details{
This function calculates the maximum distance left and right of the current cell that it can expand tangentially based on certain conditions. It checks if the current cell's radial position and radius overlap with another cell's radial position and radius. If they do, it evaluates the differentiation status and cell type of the other cell to determine if it can affect the expansion of the current cell. The function updates the maximum non-crushable distances accordingly.

code translated from Pascal to R by: ChatGPT, Annemarie Eckes-Shephard
}
\author{
David Drew
}
\keyword{CalculateTanCrushableDistance}
