% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aux_Functions_GrowCellWall.R
\name{CalculateWallThickness}
\alias{CalculateWallThickness}
\title{Calculate Wall Thickness}
\usage{
CalculateWallThickness(n)
}
\arguments{
\item{n}{The cell number index}
}
\description{
Calculates the wall thickness of cells.
}
\details{
This function calculates the wall thickness of cells based on their differentiation status.

Code translated from Pascal to R by: ChatGPT, Annemarie Eckes-Shephard
}
\author{
David Drew
}
\keyword{area}
\keyword{area,}
\keyword{cell}
\keyword{differentiation}
\keyword{lumen}
\keyword{status,}
\keyword{thickness,}
\keyword{wall}
