% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreModuleList.R
\name{CoreModuleList}
\alias{CoreModuleList}
\title{CoreModuleList}
\usage{
CoreModuleList(day, CellIdentityDeterminationMethod)
}
\value{
This function does not return any value.
}
\description{
This function performs a series of calculations and operations related to cell development and growth.
}
\details{
This function executes several subroutines to perform various tasks, including updating status information, refreshing the main form, calculating cell neighbors, calculating morphogenic fields, determining cell-specific durations, determining cells, calculating cell-specific allocation, dividing cells, calculating cell growth potential, growing cell wall, calculating cell interactions, drawing environmental graphs, and writing daily data.

Code translated from Pascal to R by: ChatGPT, Annemarie Eckes-Shephard
}
\author{
David Drew
}
\keyword{calculations,}
\keyword{cell}
\keyword{data}
\keyword{development,}
\keyword{graphs,}
\keyword{growth,}
\keyword{interactions,}
