% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aux_CalculateCellInteractions.R
\name{CrushCells}
\alias{CrushCells}
\title{Crush Cells}
\usage{
CrushCells(positioncellcounter)
}
\description{
This function crushes cells based on their type and differentiation status.
}
\details{
It checks the cell type and differentiation status of each cell and crushes the cells that meet certain conditions. The crushed cells have their cell dimensions and volumes set to zero, and their status is updated accordingly.

Code translated from Pascal to R by: ChatGPT, Annemarie Eckes-Shephard
}
\author{
David Drew
}
