% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DetermineCells.R
\name{DetermineCells}
\alias{DetermineCells}
\title{Determine Cells}
\usage{
DetermineCells(CellIdentityDeterminationMethod, TotalLoopSize, day)
}
\arguments{
\item{CellIdentityDeterminationMethod}{string with "Barlow" or "Canalisation"; hypothesis for cell identity}

\item{TotalLoopSize}{alive cells}
}
\description{
This function determines the cell identity based on the calculated auxin gradient once each day.
}
\details{
The function iterates through each cell and determines its identity based on various criteria.

code translated from Pascal to R by: ChatGPT, Annemarie Eckes-Shephard
}
\author{
David Drew
}
