% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DivideCells.R
\name{DivideCells}
\alias{DivideCells}
\title{Divide Cells}
\usage{
DivideCells(TotalLoopSize, day)
}
\value{
None
}
\description{
Loop through the entire population of cells and divide those cells that meet the requirements for mitosis and cell division.
}
\details{
This function iterates over the cell population and divides cells that meet the conditions for division based on various parameters.

Code translated from Pascal to R by: ChatGPT, Annemarie Eckes-Shephard
}
\author{
David Drew
}
\keyword{Cell}
\keyword{Division}
