% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aux_functs_daily_segment_means.R
\name{FibreSegmentMeans}
\alias{FibreSegmentMeans}
\title{FibreSegmentMeans}
\usage{
FibreSegmentMeans(day, FSegmentNumber)
}
\arguments{
\item{day}{simulation day}

\item{FSegmentNumber}{Fibre Segment Number}
}
\value{
None
}
\description{
This function calculates the average values of the fibre/vessel properties
in the segment whose sum and count was previously calculated if all cells
have died.
}
\details{
This procedure calculates the average values of the fibre/vessel properties
in the segment whose sum and count was previously calculated if all cells
have died.

Code translated from Pascal to R by: ChatGPT, Annemarie Eckes-Shephard
}
\author{
David Drew
}
\keyword{FibreSegmentMeans}
