% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aux_functions.R
\name{GetDayLength}
\alias{GetDayLength}
\title{GetDayLength}
\usage{
GetDayLength(Lat, DayOfYear)
}
\arguments{
\item{Lat}{A real number representing the latitude.}

\item{DayOfYear}{An integer representing the day of the year.}
}
\value{
The fraction of daylight hours as a real number.
}
\description{
Calculates the fraction of daylight hours based on latitude and day of the year.
}
\details{
This function calculates the fraction of daylight hours based on the latitude and day of the year. It uses mathematical formulas to determine the sunrise and sunset times.

Code translated from Pascal to R by: ChatGPT, Annemarie Eckes-Shephard
}
\author{
David Drew
}
\keyword{"day}
\keyword{"daylight}
\keyword{"latitude",}
\keyword{"sunrise",}
\keyword{"sunset"}
\keyword{hours",}
\keyword{length",}
