% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aux_functions.R
\name{GetNightHours}
\alias{GetNightHours}
\title{GetNightHours}
\usage{
GetNightHours(
  MyInputDate,
  MyLat,
  MyRainfall,
  MyMinRainfall,
  MySolarRad,
  MySolarRadYesterday,
  day
)
}
\arguments{
\item{MyInputDate}{A string representing the input date.}

\item{MyLat}{A real number representing the latitude.}

\item{MyRainfall}{A real number representing the rainfall.}

\item{MyMinRainfall}{A real number representing the minimum rainfall.}

\item{MySolarRad}{A real number representing the solar radiation.}

\item{MySolarRadYesterday}{A real number representing the solar radiation from the previous day.}

\item{day}{simulation day}
}
\value{
The number of night hours as a real number.
}
\description{
Calculates the number of night hours based on input parameters.
}
\details{
This function calculates the number of night hours based on the input date, latitude, rainfall, minimum rainfall, solar radiation, and solar radiation from the previous day.

Code translated from Pascal to R by: ChatGPT, Annemarie Eckes-Shephard
}
\author{
David Drew
}
\keyword{"day}
\keyword{"night}
\keyword{"shrinkage",}
\keyword{"sun}
\keyword{down"}
\keyword{hours",}
\keyword{length",}
