% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrowCellWall.R
\name{GrowCellWall}
\alias{GrowCellWall}
\title{Grow Cell Wall}
\usage{
GrowCellWall(day, TotalLoopSize)
}
\value{
None (the function performs calculations and updates variables)
}
\description{
The main wall thickening module.
}
\details{
This function calculates the growth of the cell wall for each cell in the loop.
It performs calculations for primary wall thickening rate, potential wall thickening rate,
secondary wall thickening rate, storage change, and wall thickness.

Code translated from Pascal to R by: ChatGPT, Annemarie Eckes-Shephard
}
\author{
David Drew
}
\keyword{cell}
\keyword{growth,}
\keyword{thickening}
\keyword{wall,}
