% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitialisModelRun.R
\name{InitialiseModelRun}
\alias{InitialiseModelRun}
\title{InitialiseModelRun}
\usage{
InitialiseModelRun()
}
\description{
Initialises the model run by setting initial values for various parameters.
}
\author{
David Drew, ChatGPT, Annemarie Eckes-Shephard
}
