% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitialiseVariables.R
\name{InitialiseVariables}
\alias{InitialiseVariables}
\title{InitialiseVariables}
\usage{
InitialiseVariables()
}
\description{
Initialises physiological parameters and variables in the model.
}
\details{
code translated from Pascal to R by: ChatGPT, Annemarie Eckes-Shephard
}
\author{
David Drew
}
