% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MorphogenicFields.R
\name{MorphogenicFields}
\alias{MorphogenicFields}
\title{Morphogenic Fields}
\usage{
MorphogenicFields(day, TotalLoopSize)
}
\arguments{
\item{day}{simulation day}

\item{TotalLoopSize}{number of alive cells}
}
\description{
This function calculates the morphogenic fields.
}
\details{
The function iterates over the total loop size and performs calculations for each cell.

Code translated from Pascal to R by: ChatGPT, Annemarie Eckes-Shephard
}
\author{
David Drew
}
