% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{params}
\alias{params}
\title{run parameters for CAMBIUM}
\format{
An object of class \code{data.frame} with 58 rows and 3 columns.
}
\source{
David Drew, file XylemParameters.xls
}
\usage{
params
}
\description{
run parameters for CAMBIUM
}
\keyword{datasets}
