% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadForcingInput.R
\name{ReadForcingInput}
\alias{ReadForcingInput}
\title{ReadForcingInput}
\usage{
ReadForcingInput(filename, StartDate = NULL, EndDate = NULL)
}
\description{
Function that reads daily forcing input into the model. [TODO] Forcings are:
wDate, ScenarioID, ASWtree, G, NPP, minmonthpredawn, greenht, ht, etacr, etaf, wf, etas, diam, tn, tx, qa, gs1, gs2, rain, stocking, vpdnow
#[TODO] this now works with the cabala spreadsheet only.
 the input header names have to be described and re-named to better reflect
 what they actually are. they are better described already in the vector
 names within the model.
 [TODO] this is not only "forcing", but also some initial values , e.g. DBH_CABALAStart
}
\details{
code translated from Pascal to R by: ChatGPT, Annemarie Eckes-Shephard
replaces ReadCabalaInput procedure in original code to make CAMBIUM more flexible and work with other daily inputs.
}
\author{
David Drew
}
