% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadParameters.R
\name{ReadParameters}
\alias{ReadParameters}
\title{ReadParameters}
\usage{
ReadParameters(filename)
}
\value{
This function "releases" the parameters into the R environment so it can be picked up by other functions.
This is a legacy to make the pascal code translation easier.
}
\description{
Read in all xylem parameters from the xylem parameters table.
}
\details{
This function reads in various xylem parameters from a table and assigns them to appropriate variables.
[TODO] can turn print statements to warning messages where appropriate

code translated from Pascal to R by: ChatGPT, Annemarie Eckes-Shephard
}
\author{
David Drew
}
\keyword{parameters}
\keyword{parameters,}
\keyword{reading}
\keyword{xylem}
