% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{runsettings}
\alias{runsettings}
\title{Run settings for CABALA model}
\format{
## `runsettings`
A data frame with 7,240 rows and 60 columns:
\describe{
  \item{Instruction_Name}{Name of the run instructions}
  \item{Instruction_Value}{Value of instruction settings. Can be string, logical, numeric, integer..}
}
}
\source{
David Drew, from the interface that worked with the Pascal Model version of CAMBIUM
}
\usage{
runsettings
}
\description{
CABALA model run settings, which the user can modify, for example to test different water snesitivity functions, or auxin channel hypotheses.
This is to be expanded by David Drew, but he will create a .csv file for this.
}
\keyword{datasets}
