% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aux_functs_daily_segment_means.R
\name{SegmentCountSumVessels}
\alias{SegmentCountSumVessels}
\title{SegmentCountSumVessels}
\usage{
SegmentCountSumVessels(CellNumber, VSegmentNumber)
}
\description{
SegmentCountSumVessels
}
