% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SumCountLoop.R
\name{SumCountLoop}
\alias{SumCountLoop}
\title{SumCountLoop}
\usage{
SumCountLoop(day)
}
\arguments{
\item{day}{simulation day}
}
\value{
None
SegmentCountSumFibres
}
\description{
This function has two main functions:
1. To make the program as fast as possible, looping is only done
   on living cells. A separate looping variable is created for this purpose,
   since even once cells die they still retain their unique identifier "cell number".
2. This function refreshes the count and summed variables of cells in different developmental stages.
}
\details{
This function updates various variables and calculates averages and counts
         related to cell properties and positions.

Code translated from Pascal to R by: ChatGPT, Annemarie Eckes-Shephard
}
\author{
David Drew
}
\keyword{Analysis,}
\keyword{Averages,}
\keyword{Cell}
\keyword{Counts}
\keyword{Developmental}
\keyword{Stages,}
