% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aux_functions.R
\name{VesselProximityCalculation}
\alias{VesselProximityCalculation}
\title{Calculate the proximity of the nearest distance and raise a "flag" if a vessel
is too close for a potential vessel to actually become a vessel}
\usage{
VesselProximityCalculation(TotalLoopSize, n)
}
\description{
Calculate the proximity of the nearest distance and raise a "flag" if a vessel
is too close for a potential vessel to actually become a vessel
}
\details{
This function calculates the proximity of the nearest distance between cells
and raises a flag if a vessel is too close to a potential vessel, indicating that it cannot become a vessel.

Code translated from Pascal to R by: ChatGPT, Annemarie Eckes-Shephard
}
\author{
David Drew
}
\keyword{calculation,}
\keyword{distance}
\keyword{flag}
\keyword{proximity,}
\keyword{vessel}
