% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aux_functs_daily_segment_means.R
\name{VesselSegmentMeans}
\alias{VesselSegmentMeans}
\title{VesselSegmentMeans}
\usage{
VesselSegmentMeans(day, VSegmentNumber)
}
\arguments{
\item{day}{simulation day}

\item{VSegmentNumber}{Vessel segment number}
}
\description{
VesselSegmentMeans
}
