% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aux_functions.R
\name{XylemDetermine_Barlow}
\alias{XylemDetermine_Barlow}
\title{Determine cell identity (fibres vs. vessels) using the concept of a
summed auxin (or other morphogenic) value (Barlow and Luck)}
\usage{
XylemDetermine_Barlow(TotalLoopSize, n, day)
}
\arguments{
\item{TotalLoopSize}{total numnber of alive cells ([TODO] I THINK..)}

\item{day}{simulation day}

\item{CellNumber}{number of cell which needs checking}
}
\value{
The cell type (FIBRE or VESSEL) based on the determined identity.
}
\description{
Determine cell identity (fibres vs. vessels) using the concept of a
summed auxin (or other morphogenic) value (Barlow and Luck)
}
\details{
This function determines the cell identity (fibres vs. vessels) based on a summed auxin score.
If the cumulative auxin score of a cell falls within the specified range, the cell has the potential to become a vessel.
The function also considers vessel proximity and cambium width to determine the final cell type.

Code translated from Pascal to R by: ChatGPT, Annemarie Eckes-Shephard
}
\author{
David Drew
}
\keyword{auxin}
\keyword{cell}
\keyword{formation,}
\keyword{identity,}
\keyword{score}
\keyword{vessel}
