% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aux_functions.R
\name{XylemDetermine_Canalisation}
\alias{XylemDetermine_Canalisation}
\title{Determine cell identity (fibres vs. vessels) using the concept of auxin canalisation
and the resultant switching of genetic programming (Sachs, 1981 etc)}
\usage{
XylemDetermine_Canalisation(TotalLoopSize, n, day)
}
\arguments{
\item{TotalLoopSize}{total numnber of alive cells ([TODO] I THINK..)}

\item{day}{simulation day}

\item{CellNumber}{number of cell which needs checking}
}
\value{
The cell type (FIBRE or VESSEL) based on the determined identity.
}
\description{
Determine cell identity (fibres vs. vessels) using the concept of auxin canalisation
and the resultant switching of genetic programming (Sachs, 1981 etc)
}
\details{
This function determines the cell identity (fibres vs. vessels) based on auxin canalisation.
If the previous auxin canalisation flag of a cell is true, the function considers vessel proximity and cambium width
to determine the final cell type.

Code translated from Pascal to R by: ChatGPT, Annemarie Eckes-Shephard
}
\author{
David Drew
}
\keyword{auxin}
\keyword{canalisation}
\keyword{cell}
\keyword{formation,}
\keyword{identity,}
\keyword{vessel}
